requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==5.2.6
django-allauth==65.13.1
fonttools==4.59.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
setuptools==65.5.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
